+++ src/app/(pages)/admin/education/history/page.tsx
@@ -11,35 +11,78 @@ import Link from "next/link";
 import Button from "@/app/components/UI/Button";
 import { useRouter } from "next/navigation";
 import Toast from "@/app/components/UI/Toast";
+import PopConfirm from "@/app/components/UI/PopConfirm";
 
-const columns: (keyof Education)[] = ["id", "name", "remark"];
+const columns: (keyof Education)[] = ["id", "education", "remark"];
 
 const Educations = () => {
   const router = useRouter();
-  const [currentPage, setCurrentPage] = useState(1);
   const [education, setEducation] = useState<Education[]>([]);
-  const [filteredEducation, setFilteredEducation] = useState<Education[]>([]);
   const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [successMessage, setSuccessMessage] = useState<string | null>(null);
+  const [isPopConfirmOpen, setIsPopConfirmOpen] = useState(false);
+  const [educationToDelete, setEducationToDelete] = useState<number | null>(null);
 
-  const rowsPerPage = 5;
-  const totalPages = Math.ceil(filteredEducation.length / rowsPerPage);
+  const [toast, setToast] = useState<{
+      message: string;
+      type: "success" | "error";
+      position: "top-right";
+    } | null>(null);
+    
+  const [pagination, setPagination] = useState({
+    totalDocs: 0,
+    totalPages: 0,
+    page: 1,
+    limit: 10,
+    pageCounter: 0,
+    hasPrevPage: false,
+    hasNextPage: false,
+  });
 
   useEffect(() => {
-    fetchEducations();
+    fetchEducations(pagination.page, pagination.limit);
   }, []);
 
-  const fetchEducations = async () => {
+
+  const searchEducations = async (searchQuery:any) => {
+    setLoading(true);
+    try {
+      const apiUrl = process.env.NEXT_PUBLIC_API_URL;
+      const response = await fetch(`${apiUrl}/api/v1/educations/all?searchQuery=${searchQuery}&page=${pagination.page}&limit=${pagination.limit}`, {
+        method: "GET",
+      });
+
+      if (response.ok) {
+        const data = await response.json();
+        setEducation(data.data);
+        setPagination(data.pagination);
+      } else {
+        console.error("Failed to fetch data");
+      }
+    } catch (error) {
+      console.error("Error fetching data:", error);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+
+  const handleSearch = (query: string) => {
+    searchEducations(query);  
+  };
+
+
+  const fetchEducations = async (page: number, limit: number) => {
     setLoading(true);
     try {
       const apiUrl = process.env.NEXT_PUBLIC_API_URL;
 
-      const response = await fetch(`${apiUrl}/educations`, { method: "GET" });
+      const response = await fetch(`${apiUrl}/api/v1/educations/all?page=${page}&limit=${limit}`, {
+        method: "GET",
+      });
       if (response.ok) {
         const data = await response.json();
-        setEducation(data);
-        setFilteredEducation(data);
+        setEducation(data.data);
+        setPagination(data.pagination);
       } else {
         console.error("Failed to fetch data");
       }
@@ -51,18 +94,33 @@ const Educations = () => {
   };
 
   const handlePageChange = (page: number) => {
-    if (page >= 1 && page <= totalPages) {
-      setCurrentPage(page);
+    setPagination((prevState) => ({
+      ...prevState,
+      page,
+    }));
+  };
+
+
+  const handleNextPage = () => {
+    if (pagination.hasNextPage) {
+      const nextPage = pagination.page + 1;
+      setPagination((prevState) => ({
+        ...prevState,
+        page: nextPage,
+      }));
+      fetchEducations(nextPage, pagination.limit);
     }
   };
 
-  const handleSearch = (query: string) => {
-    console.log("Searching for:", query);
-    const filtered = education.filter((edu) =>
-      edu.name.toLowerCase().includes(query.toLowerCase())
-    );
-    setFilteredEducation(filtered);
-    setCurrentPage(1);
+  const handlePrevPage = () => {
+    if (pagination.hasPrevPage) {
+      const prevPage = pagination.page - 1;
+      setPagination((prevState) => ({
+        ...prevState,
+        page: prevPage,
+      }));
+      fetchEducations(prevPage, pagination.limit);
+    }
   };
 
   const handleAction = (action: string, row: Record<string, any>) => {
@@ -75,31 +133,54 @@ const Educations = () => {
         router.push(`/admin/education/update/${row.id}`);
         break;
       case "delete":
-        handleDelete(row.id);
+        setEducationToDelete(row.id);
+        setIsPopConfirmOpen(true);
         break;
       default:
         break;
     }
   };
 
-  const handleDelete = async (id: number) => {
+  const handleDelete = async () => {
+    if (educationToDelete === null) return;
+
     try {
       const apiUrl = process.env.NEXT_PUBLIC_API_URL;
-      const response = await fetch(`${apiUrl}/educations/${id}`, {
+      const response = await fetch(`${apiUrl}/api/v1/educations/${educationToDelete}`, {
         method: "DELETE",
       });
 
       if (!response.ok) {
         throw new Error("Failed to delete the education");
       }
-      setError(null);
-      setSuccessMessage(`Education with id ${id} deleted successfully`);
-      console.log(`Education with id ${id} deleted successfully`);
-      fetchEducations();
+   
+      console.log(`Education with id ${educationToDelete} deleted successfully`);
+      fetchEducations(pagination.page, pagination.limit);
+      setToast({
+        message: "You have successfully deleted Education.",
+        type: "success",
+        position: "top-right",
+      });
+      setIsPopConfirmOpen(false); 
     } catch (error) {
-      setError(error);
       console.error("Error deleting the education: ", error);
+      setToast({
+        message: `${error.message}`,
+        type: "error",
+        position: "top-right",
+      });
     }
+  };
+
+
+  const handleCancel = () => {
+    setIsPopConfirmOpen(false);
+    setEducationToDelete(null);
+  };
+
+  const handleClearSearch = () => {
+    searchEducations("");  
+    fetchEducations(pagination.page, pagination.limit);
   };
 
   return (
@@ -110,11 +191,12 @@ const Educations = () => {
         borderColor="border-primary"
         borderThickness="border-b-4"
       >
-        <div className="flex flex-1 items-center justify-between m-2 w-full">
+        <div className="flex flex-1 items-center justify-between mb-2 w-full">
           <Search
             onSearch={handleSearch}
+            onClear={handleClearSearch}
             placeholder="Search Educations..."
-            buttonText="Search Educations"
+            buttonText="Search"
           />
           <Link href="/admin/education/create">
             <Button
@@ -131,38 +213,34 @@ const Educations = () => {
           <div className="ml-2 text-red-500">Loading...</div>
         ) : (
           <>
-            <Table
-              columns={columns}
-              data={filteredEducation.slice(
-                (currentPage - 1) * rowsPerPage,
-                currentPage * rowsPerPage
-              )}
-              onAction={handleAction}
-            />
+            <Table columns={columns} data={education} onAction={handleAction} />
             <div className="flex justify-end mt-4">
               <Pagination
-                currentPage={currentPage}
-                totalPages={totalPages}
-                onPageChange={handlePageChange}
+                 currentPage={pagination.page}
+                 totalPages={pagination.totalPages}
+                 onPageChange={handlePageChange}
+                 hasNextPage={pagination.hasNextPage}
+                 hasPrevPage={pagination.hasPrevPage}
+                 onNextPage={handleNextPage}
+                 onPrevPage={handlePrevPage}
               />
             </div>
           </>
         )}
+        <PopConfirm
+        isOpen={isPopConfirmOpen}
+        onConfirm={handleDelete}
+        onCancel={handleCancel}
+        message="Are you sure you want to delete this category?"
+        title="Delete Category"
+      />
       </BoxWrapper>
-      {successMessage && (
+      {toast && (
         <Toast
-          message={successMessage}
-          type="success"
-          position="top-right"
-          onClose={() => setSuccessMessage(null)}
-        />
-      )}
-      {error && (
-        <Toast
-          message={error}
-          type="error"
-          position="top-right"
-          onClose={() => setError(null)}
+          message={toast.message}
+          type={toast.type}
+          position={toast.position}
+          onClose={() => setToast(null)}
         />
       )}
     </>
