+++ src/app/(pages)/admin/source/history/page.tsx
@@ -1,39 +1,71 @@
 "use client";
 
 import React, { useEffect, useState } from "react";
-import { FaExclamationTriangle, FaPlus } from "react-icons/fa";
+import { FaInfoCircle, FaPlus } from "react-icons/fa";
 import BoxWrapper from "@/app/components/UI/BoxWrapper";
 import Table from "@/app/components/UI/Table";
 import Pagination from "@/app/components/UI/Pagination";
 import Search from "@/app/components/UI/Search";
-import { Education } from "@/app/model/EducationModel";
 import { useRouter } from "next/navigation";
 import Link from "next/link";
 import Button from "@/app/components/UI/Button";
+import { Source } from "@/app/model/Source";
+import Toast from "@/app/components/UI/Toast";
+import PopConfirm from "@/app/components/UI/PopConfirm";
 
-const columns: (keyof Education)[] = ["id", "name", "remark"];
+const columns: (keyof Source)[] = ["id", "source", "remark"];
 
 const SourceOfInformation = () => {
   const router = useRouter();
-  const [currentPage, setCurrentPage] = useState(1);
   const [sources, setSources] = useState([]);
   const [loading, setLoading] = useState(true);
-
-  const rowsPerPage = 5;
-  const totalPages = Math.ceil(sources.length / rowsPerPage);
-
-  const currentData = sources.slice(
-    (currentPage - 1) * rowsPerPage,
-    currentPage * rowsPerPage
-  );
+  const [isPopConfirmOpen, setIsPopConfirmOpen] = useState(false);
+  const [sourceToDelete, setSourceToDelete] = useState<number | null>(null);
+  const [toast, setToast] = useState<{
+    message: string;
+    type: "success" | "error";
+    position: "top-right";
+  } | null>(null);
+  const [pagination, setPagination] = useState({
+    totalDocs: 0,
+    totalPages: 0,
+    page: 1,
+    limit: 10,
+    pageCounter: 0,
+    hasPrevPage: false,
+    hasNextPage: false,
+  });
 
   useEffect(() => {
-    fetchSources();
+    fetchSources(pagination.page, pagination.limit);
   }, []);
 
   const handlePageChange = (page: number) => {
-    if (page >= 1 && page <= totalPages) {
-      setCurrentPage(page);
+    setPagination((prevState) => ({
+      ...prevState,
+      page,
+    }));
+  };
+
+  const handleNextPage = () => {
+    if (pagination.hasNextPage) {
+      const nextPage = pagination.page + 1;
+      setPagination((prevState) => ({
+        ...prevState,
+        page: nextPage,
+      }));
+      fetchSources(nextPage, pagination.limit);
+    }
+  };
+
+  const handlePrevPage = () => {
+    if (pagination.hasPrevPage) {
+      const prevPage = pagination.page - 1;
+      setPagination((prevState) => ({
+        ...prevState,
+        page: prevPage,
+      }));
+      fetchSources(prevPage, pagination.limit);
     }
   };
 
@@ -47,41 +79,84 @@ const SourceOfInformation = () => {
         router.push(`/admin/source/update/${row.id}`);
         break;
       case "delete":
-        handleDelete(row.id);
+        setSourceToDelete(row.id);
+        setIsPopConfirmOpen(true);
         break;
       default:
         break;
     }
   };
 
-  const handleDelete = async (id: number) => {
+  const handleDelete = async () => {
+    if (sourceToDelete === null) return;
+
     try {
       const apiUrl = process.env.NEXT_PUBLIC_API_URL;
-      const response = await fetch(`${apiUrl}/sources/${id}`, {
-        method: "DELETE",
-      });
+      const response = await fetch(
+        `${apiUrl}/api/v1/sources/${sourceToDelete}`,
+        {
+          method: "DELETE",
+        }
+      );
 
       if (!response.ok) {
-        throw new Error("Failed to delete the region");
+        throw new Error("Failed to delete the Source");
       }
 
-      console.log(`Region with id ${id} deleted successfully`);
-      fetchSources();
+      console.log(`Source with id ${sourceToDelete} deleted successfully`);
+      fetchSources(pagination.page, pagination.limit);
+      setToast({
+        message: "You have successfully deleted Source.",
+        type: "success",
+        position: "top-right",
+      });
+      setIsPopConfirmOpen(false);
     } catch (error) {
-      console.error("Error deleting the region: ", error);
+      console.error("Error deleting the Source: ", error);
+      setToast({
+        message: `${error.message}`,
+        type: "error",
+        position: "top-right",
+      });
     }
   };
 
-  const fetchSources = async () => {
+  const fetchSources = async (page: number, limit: number) => {
     setLoading(true);
     try {
       const apiUrl = process.env.NEXT_PUBLIC_API_URL;
 
-      const response = await fetch(`${apiUrl}/sources`, { method: "GET" });
+      const response = await fetch(
+        `${apiUrl}/api/v1/sources/all?page=${page}&limit=${limit}`,
+        { method: "GET" }
+      );
       if (response.ok) {
         const data = await response.json();
-        setSources(data);
-        // setFilteredRegions(data);
+        setSources(data.data);
+        setPagination(data.pagination);
+      } else {
+        console.error("Failed to fetch data");
+      }
+    } catch (error) {
+      console.error("Error fetching data:", error);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const searchSources = async (searchQuery: string) => {
+    setLoading(true);
+    try {
+      const apiUrl = process.env.NEXT_PUBLIC_API_URL;
+
+      const response = await fetch(
+        `${apiUrl}/api/v1/sources/all?searchQuery=${searchQuery}&page=${pagination.page}&limit=${pagination.limit}`,
+        { method: "GET" }
+      );
+      if (response.ok) {
+        const data = await response.json();
+        setSources(data.data);
+        setPagination(data.pagination);
       } else {
         console.error("Failed to fetch data");
       }
@@ -93,51 +168,84 @@ const SourceOfInformation = () => {
   };
 
   const handleSearch = (query: string) => {
-    console.log("Searching for:", query);
+    searchSources(query);
+  };
+
+  const handleClearSearch = () => {
+    searchSources("");
+    fetchSources(pagination.page, pagination.limit);
+  };
+
+  const handleCancel = () => {
+    setIsPopConfirmOpen(false);
+    setSourceToDelete(null);
   };
 
   return (
-    <BoxWrapper
-      icon={<FaExclamationTriangle />}
-      title="Source Of Information"
-      borderColor="border-primary"
-      borderThickness="border-b-4"
-    >
-      <div className="flex flex-1 items-center justify-between m-2 w-full">
-        <Search
-          onSearch={handleSearch}
-          placeholder="Search Source..."
-          buttonText="Search Source"
-        />
-
-        <Link href="/admin/source/create">
-          <Button
-            color="primary"
-            text="Create Source"
-            icon={<FaPlus />}
-            className="ml-auto"
-            size="large"
-            borderRadius={5}
+    <>
+      <BoxWrapper
+        icon={<FaInfoCircle />}
+        title="Source Of Information"
+        borderColor="border-primary"
+        borderThickness="border-b-4"
+      >
+        <div className="flex flex-1 items-center justify-between m-2 w-full">
+          <Search
+            onSearch={handleSearch}
+            onClear={handleClearSearch}
+            placeholder="Search Source..."
+            buttonText="Search Source"
           />
-        </Link>
-      </div>
-      <div className="m-2 w-full"></div>
 
-      {loading ? (
-        <div>Loading...</div>
-      ) : (
-        <>
-          <Table columns={columns} data={currentData} onAction={handleAction} />
-          <div className="flex justify-end mt-4">
-            <Pagination
-              currentPage={currentPage}
-              totalPages={totalPages}
-              onPageChange={handlePageChange}
+          <Link href="/admin/source/create">
+            <Button
+              color="primary"
+              text="Create Source"
+              icon={<FaPlus />}
+              className="ml-auto"
+              size="large"
+              borderRadius={5}
             />
-          </div>
-        </>
+          </Link>
+        </div>
+        <div className="m-2 w-full"></div>
+
+        {loading ? (
+          <div>Loading...</div>
+        ) : (
+          <>
+            <Table columns={columns} data={sources} onAction={handleAction} />
+            <div className="flex justify-end mt-4">
+              <Pagination
+                currentPage={pagination.page}
+                totalPages={pagination.totalPages}
+                onPageChange={handlePageChange}
+                hasNextPage={pagination.hasNextPage}
+                hasPrevPage={pagination.hasPrevPage}
+                onNextPage={handleNextPage}
+                onPrevPage={handlePrevPage}
+              />
+            </div>
+          </>
+        )}
+
+        <PopConfirm
+          isOpen={isPopConfirmOpen}
+          onConfirm={handleDelete}
+          onCancel={handleCancel}
+          message="Are you sure you want to delete this Source?"
+          title="Delete Source"
+        />
+      </BoxWrapper>
+      {toast && (
+        <Toast
+          message={toast.message}
+          type={toast.type}
+          position={toast.position}
+          onClose={() => setToast(null)}
+        />
       )}
-    </BoxWrapper>
+    </>
   );
 };
